+++ node/services/utils.ts
@@ -17,3 +17,40 @@ export const settings = async (context: Context): Promise<any> => {
     return null
   }
 }
+
+export const replaceDefaultSchedule = async (
+  context: Context,
+  accounts: Account[],
+  accountCode: string
+): Promise<Account[]> => {
+  const {
+    clients: { vbase },
+  } = context
+
+  const savedPayoutSchedules = await vbase.getJSON<{
+    [accountCode: string]: string
+  }>('adyen-platforms', 'updatedPayoutSchedule', true)
+
+  if (
+    !savedPayoutSchedules ||
+    !Object.prototype.hasOwnProperty.call(savedPayoutSchedules, accountCode)
+  ) {
+    return accounts
+  }
+
+  return accounts.map(account => {
+    if (
+      accountCode === account.accountCode &&
+      account.payoutSchedule?.schedule === 'DEFAULT'
+    ) {
+      return {
+        ...account,
+        payoutSchedule: {
+          schedule: savedPayoutSchedules[accountCode],
+        },
+      }
+    }
+
+    return account
+  })
+}